requirements.txt
mistralai>=1.0.0
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
rich>=13.0.0
python-docx>=1.0.0
docx2txt>=0.8
openpyxl>=3.1.0
xlrd>=2.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
